The_Six_Meals/requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.0
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-yasg==1.21.8
filelock==3.15.4
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.13.52
phonenumberslite==8.13.52
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.1.0
PyMySQL==1.1.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.31
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.26.3
Werkzeug==3.0.3
